+++ src/app/(pages)/(payment)/payment/components/(address)/DeliveryAddress .tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import DownButton from '@/components/icons/DownButton';
-import PlusIcon from '@/components/icons/PlusIcon';
 import UpButton from '@/components/icons/UpButton';
 import { AllAddresses } from '@/types/deliveryAddress';
 import useDeliveryStore from '@/zustand/payment/useDeliveryStore';
@@ -9,9 +8,9 @@ import {
   DropdownMenu,
   DropdownMenuTrigger
 } from '@radix-ui/react-dropdown-menu';
-import Link from 'next/link';
 import { SetStateAction, useEffect, useState } from 'react';
-import AddressChangeButton from './AddressChangeButton';
+import AddAddressButton from './AddAddressButton';
+import AddressSummaryCard from './AddressSummaryCard';
 
 interface Props {
   initialData: AllAddresses;
@@ -44,20 +43,6 @@ const DeliveryAddress = ({
     setAddress(selectedAddress);
   }, [selectedAddress, setAddress]);
 
-  const {
-    id,
-    addressName,
-    receiverName,
-    phoneNumber,
-    baseAddress,
-    detailAddress
-  } = selectedAddress;
-
-  const zipCode = baseAddress.match(/\((\d+)\)/)?.[1];
-  const baseAddressWithoutZipCode = baseAddress.split('(')[0];
-
-  const ADD_ADDRESS_PAGE = '/my-page/setting/delivery-address/add-new';
-
   return (
     <div className="bg-white p-4 flex flex-col gap-2 rounded-[12px] border-2 border-[#E0E0E0] mb-4">
       <div className="flex flex-col justify-between items-center">
@@ -66,33 +51,10 @@ const DeliveryAddress = ({
         </h2>
       </div>
 
-      {/* 배송지 없을 경우(추가 버튼) */}
-      {!defaultAddress ? (
-        <Link href={`${ADD_ADDRESS_PAGE}?from=payment`} className="w-full">
-          <button className="w-full flex justify-center items-center gap-2 px-4 py-3 h-10 border-[1px] border-primary-20 text-primary-20 rounded-[8px]">
-            <PlusIcon color={'#9C6D2E'} />
-            <p className="font-semibold">배송지 추가하기</p>
-          </button>
-        </Link>
+      {!selectedAddress || !defaultAddress ? (
+        <AddAddressButton />
       ) : (
-        <div className="flex flex-col gap-2">
-          <div className="flex justify-between items-center">
-            <p className="text-label-strong text-[16px] font-semibold">
-              {addressName}
-            </p>
-            {/* 변경 버튼 */}
-            <AddressChangeButton selectedAddressId={id} />
-          </div>
-          <p className="text-label-strong">{receiverName}</p>
-          <p className="text-label-alternative">{phoneNumber}</p>
-          <div className="text-label-strong">
-            <div className="flex">
-              {baseAddressWithoutZipCode}
-              {detailAddress !== '' && <p>, {detailAddress}</p>}
-            </div>
-            <p>({zipCode})</p>
-          </div>
-        </div>
+        <AddressSummaryCard selectedAddress={selectedAddress} />
       )}
 
       {/* 구분선 */}
